Convert SRMR from function to class
The new form is consistent with the structure of the other
intelligibility models in PAMBOX. The class has a `predict` method that
takes the clean, mixture, and noise only. In this case, only the mixture
is used.

The tests are also modified to use the new structure.

The `calc_erbs` and `calc_cutoffs` could possibly also be converted to
regular methods, instead of static methods.